1.完成数据局表的合并

+++ backend/controllers/ChatController.php
@@ -2,7 +2,7 @@
 
 namespace backend\controllers;
 
-use common\models\Chat;
+use backend\models\Chat;
 use common\services\Request;
 use yii\helpers\Json;
 use yii\web\Response;
@@ -13,14 +13,12 @@ class ChatController extends ObjectController
     {
         
         if (\Yii::$app->request->get('show') == 1) {
-            $data = Chat::find()->andWhere(["status" => 1])->asArray()->all();
+            $data = Chat::find()->andWhere(["useable" => 1])->asArray()->all();
         } elseif (\Yii::$app->request->get('show') == 0) {
-            $data = Chat::find()->andWhere(["status" => 0])->asArray()->all();
+            $data = Chat::find()->andWhere(["useable" => 0])->asArray()->all();
         } else {
             $data = Chat::find()->asArray()->all();
         }
-        
-        
         return $this->render('index',['data'=>$data]);
     }
     
@@ -35,14 +33,13 @@ class ChatController extends ObjectController
         if(\Yii::$app->request->isPost)
         {
             \Yii::$app->response->format = Response::FORMAT_JSON;
-            if($model->add(\Yii::$app->request->post()))
+            if($model->load(\Yii::$app->request->post()) && $model->save())
             {
                 return ['code'=>1,'message'=>'添加成功'];
             }
             $message = $model->getFirstErrors();
             $message = reset($message);
             return ['code'=>0,'message'=>$message];
-            
         }
         return $this->render('add',['model'=>$model]);
     }
@@ -59,7 +56,7 @@ class ChatController extends ObjectController
         if(\Yii::$app->request->isPost)
         {
             \Yii::$app->response->format = Response::FORMAT_JSON;
-            if($model->edit(\Yii::$app->request->post()))
+            if($model->load(\Yii::$app->request->post()) && $model->save())
             {
                 return ['code'=>1,'message'=>'修改成功'];
             }
@@ -81,17 +78,7 @@ class ChatController extends ObjectController
         \Yii::$app->response->format = Response::FORMAT_JSON;
         $id     = \Yii::$app->request->get('id');
         $model  = Chat::findOne($id);
-        $data =[];
-        $data['id']=$model->id;
-        $datas = Json::encode($data);
-        $url = \Yii::$app->params['Api'].'/control/deleteChat';
-        $re = Request::request_post_raw($url,$datas);
-        if ($re['code']==1){
-            $model->delete();
-            return ['code'=>1,'message'=>'删除成功'];
-        }
-        return ['code'=>0,'message'=>'删除失败'];
-        /*if($model)
+        if($model)
         {
             if($model->delete()) {
                 return ['code'=>1,'message'=>'删除成功'];
@@ -100,19 +87,6 @@ class ChatController extends ObjectController
             $messge = reset($messge);
             return ['code'=>0,'message'=>$messge];
         }
-        return ['code'=>0,'message'=>'删除的ID不存在'];*/
+        return ['code'=>0,'message'=>'删除的ID不存在'];
     }
-    
-    //同步数据
-    public function actionGetchat(){
-        $this->layout = false;
-        \Yii::$app->response->format = Response::FORMAT_JSON;
-        $code = Chat::GetChat();
-        if ($code ==1){
-            return ['code'=>1,'message'=>'同步成功'];
-        }
-        return ['code'=>0,'message'=>'同步失败'];
-    }
-    
-    
 }
